// openapi-defs/generated-clients/carp-ts/models/AddressUsedRequest.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Address } from './Address';
import { AddressUsedRequestAllOf } from './AddressUsedRequestAllOf';
import { BlockTxPair } from './BlockTxPair';
import { Pagination } from './Pagination';
import { HttpFile } from '../http/http';

export class AddressUsedRequest {
    'addresses': Array<Address>;
    'after'?: BlockTxPair;
    /**
    * block hash - inclusive
    */
    'untilBlock': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "addresses",
            "baseName": "addresses",
            "type": "Array<Address>",
            "format": ""
        },
        {
            "name": "after",
            "baseName": "after",
            "type": "BlockTxPair",
            "format": ""
        },
        {
            "name": "untilBlock",
            "baseName": "untilBlock",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AddressUsedRequest.attributeTypeMap;
    }

    public constructor() {
    }
}