// openapi-defs/generated-clients/carp-ts/models/AfterBlockPagination.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BlockTxPair } from './BlockTxPair';
import { HttpFile } from '../http/http';

export class AfterBlockPagination {
    'after'?: BlockTxPair;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "after",
            "baseName": "after",
            "type": "BlockTxPair",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AfterBlockPagination.attributeTypeMap;
    }

    public constructor() {
    }
}