// openapi-defs/generated-clients/carp-ts/models/BlockInfoAllOf.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class BlockInfoAllOf {
    'isValid': boolean;
    /**
    * index of tx in block
    */
    'indexInBlock': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "isValid",
            "baseName": "isValid",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "indexInBlock",
            "baseName": "indexInBlock",
            "type": "number",
            "format": "double"
        }    ];

    static getAttributeTypeMap() {
        return BlockInfoAllOf.attributeTypeMap;
    }

    public constructor() {
    }
}