// openapi-defs/generated-clients/carp-ts/models/BlockLatestResponse.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BlockSubset } from './BlockSubset';
import { HttpFile } from '../http/http';

export class BlockLatestResponse {
    'block': BlockSubset;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "block",
            "baseName": "block",
            "type": "BlockSubset",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BlockLatestResponse.attributeTypeMap;
    }

    public constructor() {
    }
}