// openapi-defs/generated-clients/carp-ts/models/BlockTxPair.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class BlockTxPair {
    /**
    * tx hash
    */
    'tx': string;
    /**
    * block hash
    */
    'block': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "tx",
            "baseName": "tx",
            "type": "string",
            "format": ""
        },
        {
            "name": "block",
            "baseName": "block",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BlockTxPair.attributeTypeMap;
    }

    public constructor() {
    }
}