// openapi-defs/generated-clients/carp-ts/models/Cip25Response.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Cip25Response {
    'cip25': { [key: string]: { [key: string]: string; }; };

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "cip25",
            "baseName": "cip25",
            "type": "{ [key: string]: { [key: string]: string; }; }",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Cip25Response.attributeTypeMap;
    }

    public constructor() {
    }
}