// openapi-defs/generated-clients/carp-ts/models/CredentialAddressRequest.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Credential } from './Credential';
import { CredentialAddressRequestAllOf } from './CredentialAddressRequestAllOf';
import { UntilBlockPagination } from './UntilBlockPagination';
import { HttpFile } from '../http/http';

export class CredentialAddressRequest {
    'after'?: string;
    'credentials': Array<Credential>;
    /**
    * block hash - inclusive
    */
    'untilBlock': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "after",
            "baseName": "after",
            "type": "string",
            "format": ""
        },
        {
            "name": "credentials",
            "baseName": "credentials",
            "type": "Array<Credential>",
            "format": ""
        },
        {
            "name": "untilBlock",
            "baseName": "untilBlock",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CredentialAddressRequest.attributeTypeMap;
    }

    public constructor() {
    }
}