// openapi-defs/generated-clients/carp-ts/models/CredentialAddressRequestAllOf.ts
/**
 * carp
 * API for the Postgres database generated by Carp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Credential } from './Credential';
import { HttpFile } from '../http/http';

export class CredentialAddressRequestAllOf {
    'after'?: string;
    'credentials': Array<Credential>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "after",
            "baseName": "after",
            "type": "string",
            "format": ""
        },
        {
            "name": "credentials",
            "baseName": "credentials",
            "type": "Array<Credential>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CredentialAddressRequestAllOf.attributeTypeMap;
    }

    public constructor() {
    }
}